// app/src/types/IDiscoverySidebarSection.ts
export type IDiscoverySidebarSection =
  | 'terms'
  | 'clusters'
  | 'domains'
  | 'pills'
  | 'verbs'
  | 'intent'
  | 'groups'
  | 'filters'
  | 'value'
  | 'ranking'
  | 'competitors'
  | 'semantic-clusters'
  | 'semantic-super-clusters'
  | 'add-keywords'
  | 'help'
  | 'recipes'
  | 'ew-patterns';